requirements.txt
python-dotenv
imap-tools
msal